% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr_mgpu}
\alias{qr_mgpu}
\title{Obtain the Q matrix from QR factorisation of a square matrix using MAGMA  multi-gpu routines}
\usage{
qr_mgpu(matrix,  printInfo=FALSE)
}
\arguments{
\item{matrix}{- the input matrix to be used in QR factorisation. It must be square}
\item{printInfo}{set to TRUE if additionaly output is needed}
}
\value{
The Q matrix is returned.
}
\description{
This function performs the QR factorisation of the input matrix and returns the Q matrix.  
The method involves the offload of the matrix data to a seperate syevd_server executable by copying data into a shared memory area and
signalling to the server that the data is availble. This function will block until the server has completed the decomposition. The
function checks that the input is square.
N.B. The maximum size allowed of the input matrix is goverend by what was provided in the MagmaQR::RunServer() function. The server
will automatically be restarted with a larger shared memory area if the user wants to perform QR on a larger matrix.
}

