% Generated by roxygen2 (4.1.0): do not edit by hand
\name{StopServer}
\alias{StopServer}
\title{Function signals to the server through shared memory region to terminate from
its main loop and then deletes the client CSharedMemory object}
\usage{
StopServer()
}
\description{
Function signals to the server through shared memory region to terminate from
its main loop and then deletes the client CSharedMemory object
}

